+++ jsonschema-csr/src/lib.rs
@@ -4,4 +4,4 @@ mod vocabularies;
 pub use compilation::JsonSchema;
 
 #[cfg(feature = "benchmark-internals")]
-pub use compilation::resolver::{scope_of, Resolver};
+pub use compilation::resolving::{scope_of, Resolver};
